more test app cleanup
