+++ code/app/build.gradle
@@ -35,12 +35,10 @@ buildscript {
 }
 
 def adobeDependencies = [
-        core       : "com.adobe.marketing.mobile:core:${mavenCoreVersion}",
-        userProfile: "com.adobe.marketing.mobile:userprofile:+",
-        edgeidentity   : "com.adobe.marketing.mobile:edgeidentity:+",
-        lifecycle  : "com.adobe.marketing.mobile:lifecycle:+",
-        signal     : "com.adobe.marketing.mobile:signal:+",
-        edge       : "com.adobe.marketing.mobile:edge:1.1.1"
+        core        : "com.adobe.marketing.mobile:core:${mavenCoreVersion}",
+        edgeidentity: "com.adobe.marketing.mobile:edgeidentity:+",
+        lifecycle   : "com.adobe.marketing.mobile:lifecycle:+",
+        edge        : "com.adobe.marketing.mobile:edge:1.1.1"
 ]
 
 dependencies {
